# perf(v2): skip runtime fileHash cache in prod & get timestamp asynchronously