--- a/packages/docusaurus-plugin-content-docs/package.json
+++ b/packages/docusaurus-plugin-content-docs/package.json
@@ -16,7 +16,7 @@
   "dependencies": {
     "@docusaurus/mdx-loader": "^2.0.0-alpha.33",
     "@docusaurus/utils": "^2.0.0-alpha.33",
-    "cross-spawn": "^7.0.1",
+    "execa": "^3.2.0",
     "fs-extra": "^8.1.0",
     "globby": "^10.0.1",
     "import-fresh": "^3.1.0",