--- a/packages/docusaurus-plugin-content-docs/src/__tests__/lastUpdate.test.ts
+++ b/packages/docusaurus-plugin-content-docs/src/__tests__/lastUpdate.test.ts
@@ -8,7 +8,6 @@
 import fs from 'fs';
 import path from 'path';
 import shell from 'shelljs';
-import spawn from 'cross-spawn';
 
 import lastUpdate from '../lastUpdate';
 
@@ -17,8 +16,8 @@ describe('lastUpdate', () => {
     __dirname,
     '__fixtures__/simple-site/docs/hello.md',
   );
-  test('existing test file in repository with Git timestamp', () => {
-    const lastUpdateData = lastUpdate(existingFilePath);
+  test('existing test file in repository with Git timestamp', async () => {
+    const lastUpdateData = await lastUpdate(existingFilePath);
     expect(lastUpdateData).not.toBeNull();
 
     const {author, timestamp} = lastUpdateData;
@@ -29,45 +28,40 @@ describe('lastUpdate', () => {
     expect(typeof timestamp).toBe('number');
   });
 
-  test('non-existing file', () => {
+  test('non-existing file', async () => {
+    const consoleMock = jest.spyOn(console, 'error');
+    consoleMock.mockImplementationOnce(err => {
+      expect(err).toMatchInlineSnapshot(
+        `[Error: Command failed with exit code 128: git log -1 --format=%ct, %an /mnt/c/Users/endij/Desktop/Linux/Docusaurus/packages/docusaurus-plugin-content-docs/src/__tests__/__fixtures__/.nonExisting]`,
+      );
+    });
     const nonExistingFilePath = path.join(
       __dirname,
       '__fixtures__',
       '.nonExisting',
     );
-    expect(lastUpdate(null)).toBeNull();
-    expect(lastUpdate(undefined)).toBeNull();
-    expect(lastUpdate(nonExistingFilePath)).toBeNull();
+    expect(await lastUpdate(nonExistingFilePath)).toBeNull();
+    expect(consoleMock).toHaveBeenCalledTimes(1);
+    expect(await lastUpdate(null)).toBeNull();
+    expect(await lastUpdate(undefined)).toBeNull();
+    consoleMock.mockRestore();
   });
 
-  test('temporary created file that has no git timestamp', () => {
+  test('temporary created file that has no git timestamp', async () => {
     const tempFilePath = path.join(__dirname, '__fixtures__', '.temp');
     fs.writeFileSync(tempFilePath, 'Lorem ipsum :)');
-    expect(lastUpdate(tempFilePath)).toBeNull();
+    expect(await lastUpdate(tempFilePath)).toBeNull();
     fs.unlinkSync(tempFilePath);
   });
 
-  test('Git does not exist', () => {
+  test('Git does not exist', async () => {
     const mock = jest.spyOn(shell, 'which').mockImplementationOnce(() => null);
     const consoleMock = jest.spyOn(console, 'warn').mockImplementation();
-    const lastUpdateData = lastUpdate(existingFilePath);
+    const lastUpdateData = await lastUpdate(existingFilePath);
     expect(lastUpdateData).toBeNull();
     expect(consoleMock).toHaveBeenLastCalledWith(
       'Sorry, the docs plugin last update options require Git.',
     );
-
-    consoleMock.mockRestore();
-    mock.mockRestore();
-  });
-
-  test('Error', () => {
-    const mock = jest.spyOn(spawn, 'sync').mockImplementationOnce(() => {
-      throw new Error('PERMISSION Error');
-    });
-    const consoleMock = jest.spyOn(console, 'error').mockImplementation();
-    const lastUpdateData = lastUpdate('/fake/path/');
-    expect(lastUpdateData).toBeNull();
-    expect(consoleMock).toHaveBeenLastCalledWith(new Error('PERMISSION Error'));
 
     consoleMock.mockRestore();
     mock.mockRestore();