## Motivation

- Skip fileHash caching on production calculation to improve build performance. The runtime cache is only useful in development.
- Obtain git timestamp for `showLastUpdateAuthor`/ `showLastUpdateTime` asynchronously instead of synchronously. 

Not sure why we did it synchronously previously, but since we are using child_process under the hood to actually do the git call, its better in async

P.S: We also use `execa` instead of `cross-spawn` because https://github.com/sindresorhus/execa has more stars (2.7k vs 600). It also strips the final newline from the output so you don't have to do stdout.trim(), Cleans up spawned processes when the parent process dies and have more descriptive errors.

### Have you read the [Contributing Guidelines on pull requests](https://github.com/facebook/docusaurus/blob/master/CONTRIBUTING.md#pull-requests)?

yes

## Test Plan

With fixtures of 1,026 docs and several runs

**Before -~89s**. 

<img width="618" alt="before 89s" src="https://user-images.githubusercontent.com/17883920/68490965-86dc3480-027c-11ea-8566-5e728d8d91f5.PNG">

**After ~~46s**
<img width="586" alt="after 46s" src="https://user-images.githubusercontent.com/17883920/68490964-86dc3480-027c-11ea-8c4a-2dceee2410a5.PNG">

The git call is indeed expensive, disabling `showUpdatedAuthor` and `showUpdatedTime` reduce first cold 'Load' time to only 4s. But well, this is still an optimization